+++ src/cmp.ndmtag.js
@@ -13,7 +13,7 @@ import renderCustomColor from './lib/custom-color';
 
 function handleConsentResult(cmp, vendorList = {}, consent = {}) {
 	let {vendorListVersion: listVersion} = vendorList;
-	let {created, vendorListVersion} = consent;
+	let {created} = consent;
 	if (configUpdates.privacyPolicy && window.location.href === configUpdates.privacyPolicy) {
 		// Do not show popup on privacy policy page
 		return;
@@ -25,10 +25,6 @@ function handleConsentResult(cmp, vendorList = {}, consent = {}) {
 	else if (!listVersion) {
 		log.debug('Could not determine vendor list version. Not showing consent tool');
 	}
-	// else if (vendorListVersion !== listVersion) {
-	// 	log.debug(`Consent found for version ${vendorListVersion}, but received vendor list version ${listVersion}. Showing consent tool`);
-	// 	cmp('showConsentTool');
-	// }
 	else {
 		let hasRequiredConsent = true;
 		let requiredPurposes = [];