+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Load ads after consent has been given
+- Update the documentation
+
+### Removed
+- Remove support for Improve tags
 
 ## [1.2.0] - 2019-06-13
 ### Added