+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Remove support for Improve tags
 
+### Fixed
+- Do not show domain name in title if locale is `de`.
+
 ## [1.2.0] - 2019-06-13
 ### Added
 - Add support for French sites