+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Load ads after consent has been given
+- Update the documentation
+
+### Removed
+- Remove support for Improve tags
+
+## [1.2.0] - 2019-06-13
+### Added
+- Add support for French sites
+
+### Changed
+- Update documentation
 
 ### Fixed
 - Use `display: none;` to close an ad and hide tags when inserting a fallback ad
@@ -34,3 +47,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.1.0]: https://github.com/nextdaymedia/cmp/tree/1.1.0
 [1.1.1]: https://github.com/nextdaymedia/cmp/tree/1.1.1
 [1.1.2-RC1]: https://github.com/nextdaymedia/cmp/tree/1.1.2-RC1
+[1.2.0]: https://github.com/nextdaymedia/cmp/tree/1.2.0